네트워크 라이브러리 수정 1. config 파일 타입, 네이밍, 클래스 로직 수정 2. 서버모듈 클래스 수정 및 기능 추가 3. 유저토근 클래스 수정 및 기능 추가 4. Acceptor 및 Connector 수정 진행 중

+++ ServerEngine/Config/ConfigEtc.cs
@@ -45,15 +45,15 @@ namespace ServerEngine.Config
     public interface IConfigPool
     {
         public string name { get; }
-        public long default_size { get; }
-        public long create_size { get; }
+        public int default_size { get; }
+        public int create_size { get; }
     }
     public class DefaultConfigPool : IConfigPool
     {
         public string name { get; }
-        public long default_size { get; }   
-        public long create_size { get; }
-        public DefaultConfigPool(string name, long default_size, long create_size)
+        public int default_size { get; }   
+        public int create_size { get; }
+        public DefaultConfigPool(string name, int default_size, int create_size)
         {
             this.name = name;
             this.default_size = default_size;
